refactor(stepping-slicer): update filename from `slicer` to `stepping-slicer` to reflect the content better
