+++ src/core/index.ts
@@ -1,4 +1,4 @@
-export { SteppingSlicer } from './slicer'
+export { SteppingSlicer } from './stepping-slicer'
 export * from './steps/steps'
 export * from './steps/input'
 export * from './steps/output'
