+++ src/core/stepping-slicer.ts
